Remove all method logics from User into Model
