+++ src/models/Model.ts
@@ -1,4 +1,4 @@
-import { AxiosPromise } from 'axios';
+import { AxiosPromise, AxiosResponse } from 'axios';
 import { Callback } from './Eventing';
 
 interface ModelAttributes<T> {
@@ -17,10 +17,52 @@ interface Events {
 	trigger(eventName: string): void;
 }
 
-export class Model<T> {
+interface HasId {
+	id: number;
+}
+
+export class Model<T extends HasId> {
 	constructor(
 		private attributes: ModelAttributes<T>,
 		private events: Events,
 		private sync: Sync<T>
 	) {}
+
+	get on() {
+		return this.events.on;
+	}
+
+	get trigger() {
+		return this.events.trigger;
+	}
+
+	get get() {
+		return this.attributes.get;
+	}
+
+	set(update: T): void {
+		this.attributes.set(update);
+		this.events.trigger('change');
+	}
+
+	fetch(): void {
+		const id = this.get('id');
+		if (typeof id !== 'number') {
+			throw new Error('Cannot fetch without an id');
+		}
+		this.sync.fetch(id).then((response: AxiosResponse): void => {
+			this.set(response.data as T);
+		});
+	}
+
+	save(): void {
+		this.sync
+			.save(this.attributes.getAll())
+			.then((response: AxiosResponse): void => {
+				this.trigger('save');
+			})
+			.catch(() => {
+				this.trigger('error');
+			});
+	}
 }
