+++ src/models/User.ts
@@ -19,42 +19,4 @@ export class User {
 	constructor(attrs: UserProps) {
 		this.attributes = new Attributes<UserProps>(attrs);
 	}
-
-	get on() {
-		return this.events.on;
-	}
-
-	get trigger() {
-		return this.events.trigger;
-	}
-
-	get get() {
-		return this.attributes.get;
-	}
-
-	set(update: UserProps): void {
-		this.attributes.set(update);
-		this.events.trigger('change');
-	}
-
-	fetch(): void {
-		const id = this.get('id');
-		if (typeof id !== 'number') {
-			throw new Error('Cannot fetch without an id');
-		}
-		this.sync.fetch(id).then((response: AxiosResponse): void => {
-			this.set(response.data);
-		});
-	}
-
-	save(): void {
-		this.sync
-			.save(this.attributes.getAll())
-			.then((response: AxiosResponse): void => {
-				this.trigger('save');
-			})
-			.catch(() => {
-				this.trigger('error');
-			});
-	}
 }
